+++ shared/hedvig-ui/index.ts
@@ -23,6 +23,13 @@ import {
   EnumDropdown,
   getTextFromEnumValue,
 } from './Dropdown/dropdown'
+import {
+  Form,
+  FormDropdown,
+  FormInput,
+  FormTextArea,
+  SubmitButton,
+} from './Form/form'
 
 export {
   Badge,
@@ -45,4 +52,9 @@ export {
   getTextFromEnumValue,
   Copyable,
   DateTimePicker,
+  Form,
+  SubmitButton,
+  FormDropdown,
+  FormInput,
+  FormTextArea,
 }
