+++ src/components/member/tabs/account-tab/AddEntryForm.tsx
@@ -1,3 +1,4 @@
+import { Form, FormDropdown, FormInput, SubmitButton } from '@hedvig-ui'
 import {
   AccountEntryInput,
   useAddAccountEntryToMemberMutation,
@@ -6,7 +7,6 @@ import { AddEntryInformation } from 'components/member/tabs/account-tab/AddEntry
 import { format } from 'date-fns'
 import { useContractMarketInfo } from 'graphql/use-get-member-contract-market-info'
 import { StandaloneMessage } from 'hedvig-ui/animations/standalone-message'
-import { Form, FormDropdown, FormInput, SubmitButton } from 'hedvig-ui/form'
 import { Spacing } from 'hedvig-ui/spacing'
 import React from 'react'
 import { FieldValues, FormProvider, useForm } from 'react-hook-form'
