+++ src/components/member/tabs/account-tab/AddMonthlyEntryForm.tsx
@@ -1,3 +1,4 @@
+import { Form, FormDropdown, FormInput, SubmitButton } from '@hedvig-ui'
 import { MonthlyEntryInput } from 'api/generated/graphql'
 import {
   getAddMonthlyEntryOptions,
@@ -5,7 +6,6 @@ import {
 } from 'graphql/use-add-monthly-entry'
 import { useContractMarketInfo } from 'graphql/use-get-member-contract-market-info'
 import { StandaloneMessage } from 'hedvig-ui/animations/standalone-message'
-import { Form, FormDropdown, FormInput, SubmitButton } from 'hedvig-ui/form'
 import React from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { FieldValues } from 'react-hook-form/dist/types/fields'
